Extract a utility class.
